Move ImageAttrs to image.ts
